Updates - (Kinda Lab 9 and A)
Bc I was focused on Algorithms work, I haven't been able to commit due to a lack of progress for a bit.  I sat down for a couple of days now and got cracking at it.  Added the create file function, updated some stuff to work with the disk and even added a mostly working (how I wanted it to) dark mode!  Wish I knew how to get the table borders to be white.